Refactor Amount handling to use double instead of BigDecimal
- Update Amount, AmountRequest, and AmountResponse to use double for value
- Modify AmountEmbeddable to store amount as double
- Add CategorySummaryProjection for summarizing category transactions
- Implement findByUserIdWithSummary method in JpaCategoryRepository

+++ app/src/main/java/com/fabiankevin/app/persistence/entities/embeddables/AmountEmbeddable.java
@@ -8,7 +8,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.util.Currency;
 import java.util.Optional;
 
@@ -19,7 +18,7 @@ import java.util.Optional;
 @Embeddable
 public class AmountEmbeddable {
     @Column(nullable = false)
-    private BigDecimal amount;
+    private double amount;
 
     @Column(nullable = false)
     private String currency;
