+++ app/src/main/java/com/fabiankevin/app/web/controllers/dtos/AmountRequest.java
@@ -3,11 +3,10 @@ package com.fabiankevin.app.web.controllers.dtos;
 import com.fabiankevin.app.models.Amount;
 import lombok.Builder;
 
-import java.math.BigDecimal;
 import java.util.Currency;
 
 @Builder
-public record AmountRequest(BigDecimal value, Currency currency) {
+public record AmountRequest(double value, Currency currency) {
     public Amount toAmount() {
         return new Amount(value, currency);
     }
