+++ app/src/main/java/com/fabiankevin/app/web/controllers/dtos/AmountResponse.java
@@ -3,11 +3,10 @@ package com.fabiankevin.app.web.controllers.dtos;
 import com.fabiankevin.app.models.Amount;
 import lombok.Builder;
 
-import java.math.BigDecimal;
 import java.util.Currency;
 
 @Builder
-public record AmountResponse(BigDecimal value, Currency currency) {
+public record AmountResponse(double value, Currency currency) {
     public static AmountResponse from(Amount amount) {
         return new AmountResponse(amount.value(), amount.currency());
     }
